plugins/dac_data_manager: Support more than single digit TX devices
Given the string we need to parse,  atoi() look like a good fit here ...

The function first discards as many whitespace characters (as in isspace)
as necessary until the first non-whitespace character is found.
Then, starting from this character, takes an optional initial plus or
minus sign followed by as many base-10 digits as possible, and interprets
them as a numerical value.
The string can contain additional characters after those that form the
integral number, which are ignored and have no effect on the behavior
of this function.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>